requirements.txt
filelock==3.16.1
fsspec==2025.3.0
jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
pandas==2.0.3
pillow==10.4.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing-extensions==4.13.2
tzdata==2025.2
